requirements.txt
guizero==1.1.0
PyAudio==0.2.11
pyserial==3.4
SpeechRecognition==3.8.1
